+++ app/src/main/AndroidManifest.xml
@@ -63,14 +63,5 @@
                 android:value="androidx.startup"
                 tools:node="remove" />
         </provider>
-
-        <!-- > Firebase push notifications <-->
-        <service
-            android:name=".core.push.firebase.FcmService"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
     </application>
 </manifest>
