--- core/push/src/main/kotlin/social/firefly/core/push/firebase/FcmService.kt
@@ -1,56 +1,0 @@
-package social.firefly.core.push.firebase
-
-import com.google.firebase.messaging.FirebaseMessagingService
-import com.google.firebase.messaging.RemoteMessage
-import kotlinx.coroutines.launch
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import org.koin.core.component.KoinComponent
-import org.koin.core.component.inject
-import social.firefly.common.appscope.AppScope
-import social.firefly.core.accounts.AccountsManager
-import social.firefly.core.push.KeyManager
-import timber.log.Timber
-
-class FcmService : FirebaseMessagingService(), KoinComponent {
-
-    private val keyManager: KeyManager by inject()
-//    private val pushRepository: PushRepository by inject()
-    private val coroutineScope: AppScope by inject()
-    private val accountsManager: AccountsManager by inject()
-
-    private val tag = FcmService::class.simpleName!!
-
-    override fun onMessageReceived(message: RemoteMessage) {
-        Timber.tag(tag).d("message received: $message")
-        //TODO decrypt message with specific account's keys and show message
-    }
-
-    override fun onNewToken(token: String) {
-        Timber.tag(tag).d("new token: $token")
-        coroutineScope.launch {
-            accountsManager.getAllAccounts().forEach {
-                val keys = keyManager.generatePushKeys()
-
-                accountsManager.updatePushKeys(
-                    mastodonAccount = it,
-                    serializedPushKeys = Json.encodeToString(keys)
-                )
-
-                Timber.tag(tag).d("keys: $keys")
-
-                try {
-                    //TODO update endpoint
-//                val webPushSubscription = pushRepository.subscribe(
-//                    endpoint = endpoint,
-//                    p256dh = keys.publicKey,
-//                    auth = keys.authSecret,
-//                )
-//                Timber.tag(tag).d("Web push subscription: $webPushSubscription")
-                } catch (e: Exception) {
-                    Timber.tag(tag).e(e)
-                }
-            }
-        }
-    }
-}
