+++ gradle/libs.versions.toml
@@ -72,8 +72,6 @@ coil = { module = "io.coil-kt:coil-compose", version.ref = "coil" }
 coil-gif = { module = "io.coil-kt:coil-gif", version.ref = "coil" }
 coil-video = { module = "io.coil-kt:coil-video", version.ref = "coil" }
 espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espresso-core" }
-firebase = { module = "com.google.firebase:firebase-bom", version = "33.3.0" }
-firebase-messaging = { module = "com.google.firebase:firebase-messaging" }
 google-truth = { module = "com.google.truth:truth", version.ref = "google-truth" }
 google-material = { module = "com.google.android.material:material", version = "1.12.0" }
 jakewharton-timber = { module = "com.jakewharton.timber:timber", version = "5.0.1" }
@@ -99,7 +97,8 @@ protobuf-kotlin-lite = { module = "com.google.protobuf:protobuf-kotlin-lite", ve
 protobuf-protoc = { module = "com.google.protobuf:protoc", version.ref = "protobuf" }
 square-okhttp = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp" }
 square-okhttp-logging = { module = "com.squareup.okhttp3:logging-interceptor", version.ref = "okhttp" }
-unifiedPush-androidConnector = { module = "com.github.UnifiedPush:android-connector", version = "2.4.0" }
+unifiedPush-androidConnector = { module = "org.unifiedpush.android:connector", version = "2.5.0" }
+unifiedPush-androidConnector-ui = { module = "org.unifiedpush.android:connector-ui", version = "1.1.0-rc2" }
 
 # Dependencies of build-logic
 android-gradlePlugin = { module = "com.android.tools.build:gradle", version.ref = "android-gradle-plugin" }
